Consolidated bug tests
